requirements.txt
mlb-statsapi>=1.9.0
pybaseball>=2.2.5
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
